// client/lib/permissions/codes.ts
export enum PermissionCode {
  EMPLOYEE_VIEW = "EMPLOYEE_VIEW",
  EMPLOYEE_CREATE = "EMPLOYEE_CREATE",
  EMPLOYEE_UPDATE = "EMPLOYEE_UPDATE",
  EMPLOYEE_DELETE = "EMPLOYEE_DELETE",

  ATTENDANCE_VIEW = "ATTENDANCE_VIEW",
  ATTENDANCE_CREATE = "ATTENDANCE_CREATE",
  ATTENDANCE_UPDATE = "ATTENDANCE_UPDATE",
  ATTENDANCE_DELETE = "ATTENDANCE_DELETE",

  LEAVE_VIEW = "LEAVE_VIEW",
  LEAVE_CREATE = "LEAVE_CREATE",
  LEAVE_UPDATE = "LEAVE_UPDATE",
  LEAVE_DELETE = "LEAVE_DELETE",

  SALARY_VIEW = "SALARY_VIEW",
  SALARY_CREATE = "SALARY_CREATE",
  SALARY_UPDATE = "SALARY_UPDATE",

  ADVANCE_VIEW = "ADVANCE_VIEW",
  ADVANCE_CREATE = "ADVANCE_CREATE",
  ADVANCE_UPDATE = "ADVANCE_UPDATE",
  ADVANCE_DELETE = "ADVANCE_DELETE",

  PAYROLL_VIEW = "PAYROLL_VIEW",
  PAYROLL_GENERATE = "PAYROLL_GENERATE",
  PAYROLL_UPDATE = "PAYROLL_UPDATE",
  PAYROLL_FINALIZE = "PAYROLL_FINALIZE",
  PAYROLL_MARK_PAID = "PAYROLL_MARK_PAID",
  PAYROLL_REOPEN_FOR_CORRECTION = "PAYROLL_REOPEN_FOR_CORRECTION",

  PAYROLL_CORRECTION_VIEW = "PAYROLL_CORRECTION_VIEW",
  PAYROLL_CORRECTION_CREATE = "PAYROLL_CORRECTION_CREATE",
  PAYROLL_CORRECTION_APPLY = "PAYROLL_CORRECTION_APPLY",
  PAYROLL_CORRECTION_REVERSE = "PAYROLL_CORRECTION_REVERSE",

  PAYSLIP_VIEW = "PAYSLIP_VIEW",
  PAYSLIP_CREATE = "PAYSLIP_CREATE",
  PAYSLIP_DOWNLOAD = "PAYSLIP_DOWNLOAD",

  USER_VIEW = "USER_VIEW",
  USER_CREATE = "USER_CREATE",
  USER_UPDATE = "USER_UPDATE",
  USER_DEACTIVATE = "USER_DEACTIVATE",

  ROLE_VIEW = "ROLE_VIEW",
  ROLE_CREATE = "ROLE_CREATE",
  ROLE_UPDATE = "ROLE_UPDATE",
  ROLE_DELETE = "ROLE_DELETE",

  COMPANY_SETTINGS_VIEW = "COMPANY_SETTINGS_VIEW",
  COMPANY_SETTINGS_UPDATE = "COMPANY_SETTINGS_UPDATE",

  DEPARTMENT_VIEW = "DEPARTMENT_VIEW",
  DEPARTMENT_CREATE = "DEPARTMENT_CREATE",
  DEPARTMENT_UPDATE = "DEPARTMENT_UPDATE",
  DEPARTMENT_DELETE = "DEPARTMENT_DELETE",

  DESIGNATION_VIEW = "DESIGNATION_VIEW",
  DESIGNATION_CREATE = "DESIGNATION_CREATE",
  DESIGNATION_UPDATE = "DESIGNATION_UPDATE",
  DESIGNATION_DELETE = "DESIGNATION_DELETE",

  SUBSCRIPTION_VIEW = "SUBSCRIPTION_VIEW",
  SUBSCRIPTION_MANAGE = "SUBSCRIPTION_MANAGE",

  AUDIT_LOG_VIEW = "AUDIT_LOG_VIEW",
}